+++ scripts/publish/src/getPackages.js
@@ -1,0 +1,6 @@
+const { exec } = require('shelljs');
+
+module.exports = () =>
+  exec('npx lerna ls --parseable', { silent: true })
+    .stdout.split('\n')
+    .filter(p => p);
