+++ scripts/publish/src/publish.js
